requirements.txt
collections-extended==1.0.2
ruamel.appconfig==0.5.4
ruamel.std.argparse==0.8.1
ruamel.yaml==0.16.0
ruamel.yaml.clib==0.1.0
yaml-1.3==0.1.0
argparse==1.4.0
setuptools==40.8.0
